fix(runtime): region chunk_idx u16→u32 —— 修全量 build 堆损坏（自举墙真凶）
`Region<T>`（GC 对象/数组区分配器）的 chunk 索引是 u16，CHUNK_SIZE=256 →
region bump 过 65535 个 chunk（≈16.7M slot）后 `next_bump.0 + 1` 溢出 u16，
release（无 overflow 检查）静默回绕到 0 → `next_bump = (0,0)` → 新分配覆盖
chunk-0 的存活对象 → **非确定堆损坏**。

现象：格式-bump PR 的 CI 两代自举 Gen0 stdlib build 到 `[24/24] z42.net`
算 sidecar BLID 时 `Std.Crypto.Blake3._compress: array index N out of bounds`
/ `ArrayGet index got Str("this")`（乱 index/len/register，每次不同）。24-lib
全量 build 已涨到逼近 u16 chunk 上限；main 正常 build 贴边未爆（绿），两代
自举 Gen0 略多而越界。stdlib 再涨正常 build 也会爆——这是真 bug 非仅两代自举。

定位：release + `-C debug-assertions` VM 跑 Gen0 精确捕获
`gc/region.rs:320: attempt to add with overflow`。

修复：chunk_idx u16→u32（`RegionEntry.location` / `RegionHandle.chunk_idx` /
`next_bump` / `free_list` / `young_list` / card-dirty / `Violation` 枚举 /
`validate`；sentinel `(u16::MAX,_)`→`(u32::MAX,_)`；arc_heap card-dirty
`u16::MAX`→`u32::MAX`）。entry_idx 保持 u16（CHUNK_SIZE≤65536）。`GcRef` 打包
的是 entry 指针非 chunk_idx → 8B handle 不受影响；外部已 `as u32`。

验证：修后 VM 跑两代自举 Gen0 全量 build **完成**（z42.net + Blake3 sidecar
正确产出，无损坏无 OOM → 16.7M 是真峰值非泄漏）；`cargo test --lib` 970/0 +
gc:: 287/0 全绿。

+++ src/runtime/src/gc/arc_heap.rs
@@ -1078,7 +1078,7 @@ impl ArcMagrGC {
                 // SAFETY: entry pointer valid for GcRef lifetime.
                 let entry = unsafe { entry_ptr.as_ref() };
                 let (ci, _) = entry.location;
-                if ci != u16::MAX {
+                if ci != u32::MAX {
                     self.region_object.lock().mark_card_dirty(ci);
                 }
             }
@@ -1087,7 +1087,7 @@ impl ArcMagrGC {
                 let entry_ptr = gc.entry_ptr();
                 let entry = unsafe { entry_ptr.as_ref() };
                 let (ci, _) = entry.location;
-                if ci != u16::MAX {
+                if ci != u32::MAX {
                     self.region_array.lock().mark_card_dirty(ci);
                 }
             }
